Fix content pack decode failure: change payload.target from Int? to Double?
Fraction templates (fras-*, frau-*) use decimal target values like 0.3
and 0.625. Swift's JSONDecoder rejects the entire file when it encounters
a Double for an Int? field, causing the "data isn't in correct format"
error and empty fallback. Changed target to Double? and added Int()
conversions at all usage sites.

+++ MathQuestKids/Content/ContentPack.swift
@@ -72,7 +72,7 @@ struct ItemPayload: Codable, Equatable {
     let right: Int?
     let minuend: Int?
     let subtrahend: Int?
-    let target: Int?
+    let target: Double?
     let tens: Int?
     let ones: Int?
 
@@ -119,7 +119,7 @@ struct ItemPayload: Codable, Equatable {
         right: Int? = nil,
         minuend: Int? = nil,
         subtrahend: Int? = nil,
-        target: Int? = nil,
+        target: Double? = nil,
         tens: Int? = nil,
         ones: Int? = nil,
         multiplicand: Int? = nil,
